fix compile (nw)

+++ src/mame/includes/metalmx.h
@@ -66,7 +66,7 @@ private:
 
 	required_shared_ptr<uint32_t> m_adsp_internal_program_ram;
 	required_shared_ptr<uint16_t> m_gsp_dram;
-	required_shared_ptr<uint16_t> m_gsp_vram;
+	required_shared_ptr<uint32_t> m_gsp_vram;
 };
 
 #endif // MAME_INCLUDES_METALMX_H
